+++ src/components/List.tsx
@@ -1,21 +1,31 @@
 import React from 'react';
+import { useDeleteTodoMutation } from '../generated/graphql';
 
 interface Props {
-  id: string | undefined;
-  title: string | undefined;
-  onDelete: (id: string | undefined) => void;
+  id: string;
+  title: string;
 }
 
 export const List: React.FC<Props> = props => {
-  // DELETEボタンをクリックすることで発火するメソッド
-  const handleClick = () => {
-    props.onDelete(props.id);
-  };
+  const [deleteTodo, { error }] = useDeleteTodoMutation();
+
+  if (error) {
+    return <div>削除出来ませんでした。</div>;
+  }
 
   return (
     <div className="item-list">
       <p className="item-name">{props.title}</p>
-      <button className="delete-btn" onClick={handleClick}>
+      <button
+        className="delete-btn"
+        onClick={() => {
+          deleteTodo({
+            variables: {
+              id: props.id
+            }
+          });
+        }}
+      >
         DELETE
       </button>
     </div>
